Improved testing setup (#441)
