+++ packages/agreement-process/test/agreementService.integration.test.ts
@@ -2,16 +2,7 @@
 /* eslint-disable functional/immutable-data */
 
 import { fail } from "assert";
-import {
-  AgreementCollection,
-  EServiceCollection,
-  FileManager,
-  ReadModelRepository,
-  TenantCollection,
-  initDB,
-  initFileManager,
-  genericLogger,
-} from "pagopa-interop-commons";
+import { genericLogger } from "pagopa-interop-commons";
 import {
   getMockAgreement,
   getMockCertifiedTenantAttribute,
@@ -24,12 +15,6 @@ import {
   getRandomAuthData,
   randomArrayItem,
   decodeProtobufPayload,
-  TEST_MONGO_DB_PORT,
-  TEST_POSTGRES_DB_PORT,
-  mongoDBContainer,
-  postgreSQLContainer,
-  minioContainer,
-  TEST_MINIO_PORT,
 } from "pagopa-interop-commons-test";
 import {
   Agreement,
@@ -51,17 +36,7 @@ import {
   generateId,
   unsafeBrandId,
 } from "pagopa-interop-models";
-import { IDatabase } from "pg-promise";
-import { StartedTestContainer } from "testcontainers";
-import {
-  afterAll,
-  afterEach,
-  beforeEach,
-  beforeAll,
-  describe,
-  expect,
-  it,
-} from "vitest";
+import { beforeEach, describe, expect, it } from "vitest";
 import {
   agreementAlreadyExists,
   agreementNotFound,
@@ -73,50 +48,18 @@ import {
 } from "../src/model/domain/errors.js";
 import { toAgreementStateV2 } from "../src/model/domain/toEvent.js";
 import { ApiAgreementPayload } from "../src/model/types.js";
-import {
-  AgreementService,
-  agreementServiceBuilder,
-} from "../src/services/agreementService.js";
-import { agreementQueryBuilder } from "../src/services/readmodel/agreementQuery.js";
-import { attributeQueryBuilder } from "../src/services/readmodel/attributeQuery.js";
-import { eserviceQueryBuilder } from "../src/services/readmodel/eserviceQuery.js";
-import {
-  ReadModelService,
-  readModelServiceBuilder,
-} from "../src/services/readmodel/readModelService.js";
-import { tenantQueryBuilder } from "../src/services/readmodel/tenantQuery.js";
-import { config } from "../src/utilities/config.js";
 import { agreementCreationConflictingStates } from "../src/model/domain/validators.js";
 import {
   CompactEService,
   CompactOrganization,
 } from "../src/model/domain/models.js";
 import {
+  readLastAgreementEvent,
   addOneAgreement,
   addOneEService,
   addOneTenant,
-  readLastAgreementEvent,
+  agreementService,
 } from "./utils.js";
-import { testArchiveAgreement } from "./testArchiveAgreement.js";
-import { testDeleteAgreement } from "./testDeleteAgreement.js";
-import { testAgreementConsumerDocuments } from "./testAgreementConsumerDocuments.js";
-import { testRejectAgreement } from "./testRejectAgreement.js";
-import { testUpdateAgreement } from "./testUpdateAgreement.js";
-import { testUpgradeAgreement } from "./testUpgradeAgreement.js";
-import { testCloneAgreement } from "./testCloneAgreement.js";
-import { testSuspendAgreement } from "./testSuspendAgreement.js";
-
-export let agreements: AgreementCollection;
-export let eservices: EServiceCollection;
-export let tenants: TenantCollection;
-export let readModelService: ReadModelService;
-export let agreementService: AgreementService;
-export let postgresDB: IDatabase<unknown>;
-export let startedPostgreSqlContainer: StartedTestContainer;
-export let startedMongodbContainer: StartedTestContainer;
-export let startedMinioContainer: StartedTestContainer;
-export let fileManager: FileManager;
-const s3OriginalBucket = config.s3Bucket;
 
 /**
  * Executes the generic agreement expectation for agreement creation process,
@@ -141,7 +84,7 @@ const expectedAgreementCreation = async (
     fail("Unhandled error: returned agreementId is undefined");
   }
 
-  const writtenEvent = await readLastAgreementEvent(agreementId, postgresDB);
+  const writtenEvent = await readLastAgreementEvent(agreementId);
 
   if (!writtenEvent) {
     fail("Creation fails: agreement not found in event-store");
@@ -184,72 +127,6 @@ const expectedAgreementCreation = async (
   return actualAgreement;
 };
 
-beforeAll(async () => {
-  startedPostgreSqlContainer = await postgreSQLContainer(config).start();
-  startedMongodbContainer = await mongoDBContainer(config).start();
-  startedMinioContainer = await minioContainer(config).start();
-
-  config.eventStoreDbPort = startedPostgreSqlContainer.getMappedPort(
-    TEST_POSTGRES_DB_PORT
-  );
-  config.readModelDbPort =
-    startedMongodbContainer.getMappedPort(TEST_MONGO_DB_PORT);
-  config.s3ServerPort = startedMinioContainer.getMappedPort(TEST_MINIO_PORT);
-
-  const readModelRepository = ReadModelRepository.init(config);
-  agreements = readModelRepository.agreements;
-  eservices = readModelRepository.eservices;
-  tenants = readModelRepository.tenants;
-
-  readModelService = readModelServiceBuilder(readModelRepository);
-  const eserviceQuery = eserviceQueryBuilder(readModelService);
-  const agreementQuery = agreementQueryBuilder(readModelService);
-  const tenantQuery = tenantQueryBuilder(readModelService);
-  const attributeQuery = attributeQueryBuilder(readModelService);
-
-  postgresDB = initDB({
-    username: config.eventStoreDbUsername,
-    password: config.eventStoreDbPassword,
-    host: config.eventStoreDbHost,
-    port: config.eventStoreDbPort,
-    database: config.eventStoreDbName,
-    schema: config.eventStoreDbSchema,
-    useSSL: config.eventStoreDbUseSSL,
-  });
-
-  if (!postgresDB) {
-    genericLogger.error("postgresDB is undefined!!");
-  }
-
-  fileManager = initFileManager(config);
-  agreementService = agreementServiceBuilder(
-    postgresDB,
-    agreementQuery,
-    tenantQuery,
-    eserviceQuery,
-    attributeQuery,
-    fileManager
-  );
-});
-
-afterEach(async () => {
-  await agreements.deleteMany({});
-  await eservices.deleteMany({});
-  await tenants.deleteMany({});
-
-  await postgresDB.none("TRUNCATE TABLE agreement.events RESTART IDENTITY");
-  await postgresDB.none("TRUNCATE TABLE catalog.events RESTART IDENTITY");
-
-  // Some tests change the bucket name, so we need to reset it
-  config.s3Bucket = s3OriginalBucket;
-});
-
-afterAll(async () => {
-  await startedPostgreSqlContainer.stop();
-  await startedMongodbContainer.stop();
-  await startedMinioContainer.stop();
-});
-
 describe("Agreement service", () => {
   describe("create agreement", () => {
     it("should succeed when EService Producer and Agreement Consumer are the same, even on unmet attributes", async () => {
@@ -266,8 +143,8 @@ describe("Agreement service", () => {
       ]);
       const tenant = getMockTenant(authData.organizationId);
 
-      await addOneEService(eservice, eservices);
-      await addOneTenant(tenant, tenants);
+      await addOneEService(eservice);
+      await addOneTenant(tenant);
 
       const agreementData: ApiAgreementPayload = {
         eserviceId,
@@ -323,9 +200,9 @@ describe("Agreement service", () => {
         [descriptor]
       );
 
-      await addOneTenant(eserviceProducer, tenants);
-      await addOneTenant(consumer, tenants);
-      await addOneEService(eservice, eservices);
+      await addOneTenant(eserviceProducer);
+      await addOneTenant(consumer);
+      await addOneEService(eservice);
 
       const apiAgreementPayload: ApiAgreementPayload = {
         eserviceId: eservice.id,
@@ -359,9 +236,9 @@ describe("Agreement service", () => {
         [descriptor]
       );
 
-      await addOneTenant(eserviceProducer, tenants);
-      await addOneTenant(consumer, tenants);
-      await addOneEService(eservice, eservices);
+      await addOneTenant(eserviceProducer);
+      await addOneTenant(consumer);
+      await addOneEService(eservice);
 
       const authData = getRandomAuthData(consumer.id); // different from eserviceProducer
       const apiAgreementPayload: ApiAgreementPayload = {
@@ -405,8 +282,8 @@ describe("Agreement service", () => {
         descriptor2,
       ]);
 
-      await addOneTenant(tenant, tenants);
-      await addOneEService(eservice, eservices);
+      await addOneTenant(tenant);
+      await addOneEService(eservice);
 
       const authData = getRandomAuthData(tenant.id);
       const apiAgreementPayload: ApiAgreementPayload = {
@@ -446,9 +323,9 @@ describe("Agreement service", () => {
         )
       );
 
-      await addOneTenant(tenant, tenants);
-      await addOneEService(eservice, eservices);
-      await addOneAgreement(otherAgreement, postgresDB, agreements);
+      await addOneTenant(tenant);
+      await addOneEService(eservice);
+      await addOneAgreement(otherAgreement);
 
       const authData = getRandomAuthData(tenant.id);
       const apiAgreementPayload: ApiAgreementPayload = {
@@ -498,7 +375,7 @@ describe("Agreement service", () => {
 
       const eservice = getMockEService(eserviceId, authData.organizationId, []);
 
-      await addOneEService(eservice, eservices);
+      await addOneEService(eservice);
 
       const apiAgreementPayload: ApiAgreementPayload = {
         eserviceId,
@@ -542,8 +419,8 @@ describe("Agreement service", () => {
         descriptor1,
       ]);
 
-      await addOneEService(eservice, eservices);
-      await addOneTenant(getMockTenant(authData.organizationId), tenants);
+      await addOneEService(eservice);
+      await addOneTenant(getMockTenant(authData.organizationId));
 
       const apiAgreementPayload: ApiAgreementPayload = {
         eserviceId,
@@ -584,8 +461,8 @@ describe("Agreement service", () => {
         descriptor,
       ]);
 
-      await addOneEService(eservice, eservices);
-      await addOneTenant(getMockTenant(authData.organizationId), tenants);
+      await addOneEService(eservice);
+      await addOneTenant(getMockTenant(authData.organizationId));
 
       const apiAgreementPayload: ApiAgreementPayload = {
         eserviceId,
@@ -622,9 +499,9 @@ describe("Agreement service", () => {
         ),
       };
 
-      await addOneTenant(consumer, tenants);
-      await addOneEService(eservice, eservices);
-      await addOneAgreement(conflictingAgreement, postgresDB, agreements);
+      await addOneTenant(consumer);
+      await addOneEService(eservice);
+      await addOneAgreement(conflictingAgreement);
 
       const authData = getRandomAuthData(consumer.id);
       const apiAgreementPayload: ApiAgreementPayload = {
@@ -652,7 +529,7 @@ describe("Agreement service", () => {
         [descriptor]
       );
 
-      await addOneEService(eservice, eservices);
+      await addOneEService(eservice);
 
       const authData = getRandomAuthData(consumer.id);
       const apiAgreementPayload: ApiAgreementPayload = {
@@ -701,9 +578,9 @@ describe("Agreement service", () => {
         [descriptor]
       );
 
-      await addOneTenant(eserviceProducer, tenants);
-      await addOneTenant(consumer, tenants);
-      await addOneEService(eservice, eservices);
+      await addOneTenant(eserviceProducer);
+      await addOneTenant(consumer);
+      await addOneEService(eservice);
 
       const authData = getRandomAuthData(consumer.id);
       const apiAgreementPayload: ApiAgreementPayload = {
@@ -760,9 +637,9 @@ describe("Agreement service", () => {
         attributes: [certifiedTenantAttribute1, certifiedTenantAttribute2],
       };
 
-      await addOneTenant(eserviceProducer, tenants);
-      await addOneTenant(consumer, tenants);
-      await addOneEService(eservice, eservices);
+      await addOneTenant(eserviceProducer);
+      await addOneTenant(consumer);
+      await addOneEService(eservice);
 
       const authData = getRandomAuthData(consumer.id);
       const apiAgreementPayload: ApiAgreementPayload = {
@@ -852,12 +729,12 @@ describe("Agreement service", () => {
         name: "EService3", // Adding name because results are sorted by esevice name
       };
 
-      await addOneTenant(tenant1, tenants);
-      await addOneTenant(tenant2, tenants);
-      await addOneTenant(tenant3, tenants);
-      await addOneEService(eservice1, eservices);
-      await addOneEService(eservice2, eservices);
-      await addOneEService(eservice3, eservices);
+      await addOneTenant(tenant1);
+      await addOneTenant(tenant2);
+      await addOneTenant(tenant3);
+      await addOneEService(eservice1);
+      await addOneEService(eservice2);
+      await addOneEService(eservice3);
 
       attribute1 = { id: generateId() };
       attribute2 = { id: generateId() };
@@ -908,12 +785,12 @@ describe("Agreement service", () => {
         producerId: eservice3.producerId,
       };
 
-      await addOneAgreement(agreement1, postgresDB, agreements);
-      await addOneAgreement(agreement2, postgresDB, agreements);
-      await addOneAgreement(agreement3, postgresDB, agreements);
-      await addOneAgreement(agreement4, postgresDB, agreements);
-      await addOneAgreement(agreement5, postgresDB, agreements);
-      await addOneAgreement(agreement6, postgresDB, agreements);
+      await addOneAgreement(agreement1);
+      await addOneAgreement(agreement2);
+      await addOneAgreement(agreement3);
+      await addOneAgreement(agreement4);
+      await addOneAgreement(agreement5);
+      await addOneAgreement(agreement6);
     });
 
     it("should get all agreements if no filters are provided", async () => {
@@ -1274,8 +1151,8 @@ describe("Agreement service", () => {
   describe("get agreement", () => {
     it("should get an agreement", async () => {
       const agreement: Agreement = getMockAgreement();
-      await addOneAgreement(agreement, postgresDB, agreements);
-      await addOneAgreement(getMockAgreement(), postgresDB, agreements);
+      await addOneAgreement(agreement);
+      await addOneAgreement(getMockAgreement());
 
       const result = await agreementService.getAgreementById(
         agreement.id,
@@ -1287,7 +1164,7 @@ describe("Agreement service", () => {
     it("should throw an agreementNotFound error when the agreement does not exist", async () => {
       const agreementId = generateId<AgreementId>();
 
-      await addOneAgreement(getMockAgreement(), postgresDB, agreements);
+      await addOneAgreement(getMockAgreement());
 
       await expect(
         agreementService.getAgreementById(agreementId, genericLogger)
@@ -1315,12 +1192,12 @@ describe("Agreement service", () => {
       tenant5 = { ...getMockTenant(), name: "Tenant 5 BazBar" };
       tenant6 = { ...getMockTenant(), name: "Tenant 6 BazFoo" };
 
-      await addOneTenant(tenant1, tenants);
-      await addOneTenant(tenant2, tenants);
-      await addOneTenant(tenant3, tenants);
-      await addOneTenant(tenant4, tenants);
-      await addOneTenant(tenant5, tenants);
-      await addOneTenant(tenant6, tenants);
+      await addOneTenant(tenant1);
+      await addOneTenant(tenant2);
+      await addOneTenant(tenant3);
+      await addOneTenant(tenant4);
+      await addOneTenant(tenant5);
+      await addOneTenant(tenant6);
 
       const agreement1 = {
         ...getMockAgreement(),
@@ -1352,11 +1229,11 @@ describe("Agreement service", () => {
         consumerId: tenant6.id,
       };
 
-      await addOneAgreement(agreement1, postgresDB, agreements);
-      await addOneAgreement(agreement2, postgresDB, agreements);
-      await addOneAgreement(agreement3, postgresDB, agreements);
-      await addOneAgreement(agreement4, postgresDB, agreements);
-      await addOneAgreement(agreement5, postgresDB, agreements);
+      await addOneAgreement(agreement1);
+      await addOneAgreement(agreement2);
+      await addOneAgreement(agreement3);
+      await addOneAgreement(agreement4);
+      await addOneAgreement(agreement5);
     });
     describe("get agreement consumers", () => {
       it("should get all agreement consumers", async () => {
@@ -1569,12 +1446,12 @@ describe("Agreement service", () => {
         name: "EService 3 FooBar",
       };
 
-      await addOneTenant(tenant1, tenants);
-      await addOneTenant(tenant2, tenants);
-      await addOneTenant(tenant3, tenants);
-      await addOneEService(eservice1, eservices);
-      await addOneEService(eservice2, eservices);
-      await addOneEService(eservice3, eservices);
+      await addOneTenant(tenant1);
+      await addOneTenant(tenant2);
+      await addOneTenant(tenant3);
+      await addOneEService(eservice1);
+      await addOneEService(eservice2);
+      await addOneEService(eservice3);
 
       const agreement1 = {
         ...getMockAgreement(eservice1.id),
@@ -1596,9 +1473,9 @@ describe("Agreement service", () => {
         state: agreementState.pending,
       };
 
-      await addOneAgreement(agreement1, postgresDB, agreements);
-      await addOneAgreement(agreement2, postgresDB, agreements);
-      await addOneAgreement(agreement3, postgresDB, agreements);
+      await addOneAgreement(agreement1);
+      await addOneAgreement(agreement2);
+      await addOneAgreement(agreement3);
     });
 
     it("should get all agreement eservices", async () => {
@@ -1824,13 +1701,4 @@ describe("Agreement service", () => {
       });
     });
   });
-
-  testAgreementConsumerDocuments();
-  testArchiveAgreement();
-  testCloneAgreement();
-  testDeleteAgreement();
-  testRejectAgreement();
-  testSuspendAgreement();
-  testUpdateAgreement();
-  testUpgradeAgreement();
 });
