+++ packages/agreement-process/vitest.config.ts
@@ -2,8 +2,10 @@ import { defineConfig } from "vitest/config";
 
 export default defineConfig({
   test: {
-    setupFiles: ["dotenv/config"],
+    globalSetup: ["./test/vitestGlobalSetup.ts"],
     testTimeout: 60000,
     hookTimeout: 60000,
+    fileParallelism: false,
+    pool: "forks"
   },
 });
