+++ packages/agreement-readmodel-writer/package.json
@@ -30,7 +30,7 @@
     "ts-node": "10.9.2",
     "typescript": "5.1.3",
     "uuid": "9.0.1",
-    "vitest": "0.33.0"
+    "vitest": "1.5.2"
   },
   "dependencies": {
     "@protobuf-ts/runtime": "2.9.4",
