+++ packages/attribute-registry-process/test/utils.ts
@@ -1,3 +1,5 @@
+import { setupTestContainersVitest } from "pagopa-interop-commons-test/index.js";
+import { afterEach, inject } from "vitest";
 import {
   Attribute,
   AttributeEvent,
@@ -8,12 +10,6 @@ import {
   toAttributeV1,
   toReadModelAttribute,
 } from "pagopa-interop-models";
-import { IDatabase } from "pg-promise";
-import {
-  AttributeCollection,
-  AuthData,
-  TenantCollection,
-} from "pagopa-interop-commons";
 import {
   ReadEvent,
   StoredEvent,
@@ -21,6 +17,57 @@ import {
   writeInEventstore,
   writeInReadmodel,
 } from "pagopa-interop-commons-test/index.js";
+import { AuthData } from "pagopa-interop-commons";
+import { readModelServiceBuilder } from "../src/services/readModelService.js";
+import { attributeRegistryServiceBuilder } from "../src/services/attributeRegistryService.js";
+
+export const { readModelRepository, postgresDB, fileManager, cleanup } =
+  setupTestContainersVitest(inject("config"));
+
+afterEach(cleanup);
+
+export const agreements = readModelRepository.agreements;
+export const eservices = readModelRepository.eservices;
+export const tenants = readModelRepository.tenants;
+export const attributes = readModelRepository.attributes;
+
+export const readModelService = readModelServiceBuilder(readModelRepository);
+export const attributeRegistryService = attributeRegistryServiceBuilder(
+  postgresDB,
+  readModelService
+);
+
+export const writeAttributeInEventstore = async (
+  attribute: Attribute
+): Promise<void> => {
+  const attributeEvent: AttributeEvent = {
+    type: "AttributeAdded",
+    event_version: 1,
+    data: { attribute: toAttributeV1(attribute) },
+  };
+  const eventToWrite: StoredEvent<AttributeEvent> = {
+    // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+    stream_id: attributeEvent.data.attribute!.id,
+    version: 0,
+    event: attributeEvent,
+  };
+
+  await writeInEventstore(eventToWrite, "attribute", postgresDB);
+};
+
+export const addOneAttribute = async (attribute: Attribute): Promise<void> => {
+  await writeAttributeInEventstore(attribute);
+  await writeInReadmodel(toReadModelAttribute(attribute), attributes);
+};
+
+export const addOneTenant = async (tenant: Tenant): Promise<void> => {
+  await writeInReadmodel(tenant, tenants);
+};
+
+export const readLastAttributeEvent = async (
+  attributeId: AttributeId
+): Promise<ReadEvent<AttributeEvent>> =>
+  await readLastEventByStreamId(attributeId, "attribute", postgresDB);
 
 export const getMockTenant = (): Tenant => ({
   name: "tenant_Name",
@@ -45,44 +92,3 @@ export const getMockAuthData = (organizationId?: TenantId): AuthData => ({
   },
   selfcareId: generateId(),
 });
-
-export const writeAttributeInEventstore = async (
-  attribute: Attribute,
-  postgresDB: IDatabase<unknown>
-): Promise<void> => {
-  const attributeEvent: AttributeEvent = {
-    type: "AttributeAdded",
-    event_version: 1,
-    data: { attribute: toAttributeV1(attribute) },
-  };
-  const eventToWrite: StoredEvent<AttributeEvent> = {
-    // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-    stream_id: attributeEvent.data.attribute!.id,
-    version: 0,
-    event: attributeEvent,
-  };
-
-  await writeInEventstore(eventToWrite, "attribute", postgresDB);
-};
-
-export const addOneAttribute = async (
-  attribute: Attribute,
-  postgresDB: IDatabase<unknown>,
-  attributes: AttributeCollection
-): Promise<void> => {
-  await writeAttributeInEventstore(attribute, postgresDB);
-  await writeInReadmodel(toReadModelAttribute(attribute), attributes);
-};
-
-export const addOneTenant = async (
-  tenant: Tenant,
-  tenants: TenantCollection
-): Promise<void> => {
-  await writeInReadmodel(tenant, tenants);
-};
-
-export const readLastAttributeEvent = async (
-  attributeId: AttributeId,
-  postgresDB: IDatabase<unknown>
-): Promise<ReadEvent<AttributeEvent>> =>
-  await readLastEventByStreamId(attributeId, "attribute", postgresDB);
