+++ packages/attribute-registry-process/test/vitestGlobalSetup.ts
@@ -1,0 +1,3 @@
+import { setupTestContainersVitestGlobal } from "pagopa-interop-commons-test";
+
+export default setupTestContainersVitestGlobal();
