+++ packages/attribute-registry-readmodel-writer/package.json
@@ -27,7 +27,7 @@
     "testcontainers": "10.2.2",
     "ts-node": "10.9.2",
     "typescript": "5.1.3",
-    "vitest": "0.33.0"
+    "vitest": "1.5.2"
   },
   "dependencies": {
     "@jm18457/kafkajs-msk-iam-authentication-mechanism": "3.1.2",
