+++ packages/attribute-registry-readmodel-writer/vitest.config.ts
@@ -2,8 +2,8 @@ import { defineConfig } from "vitest/config";
 
 export default defineConfig({
   test: {
-    setupFiles: ["dotenv/config"],
     testTimeout: 60000,
     hookTimeout: 60000,
+    setupFiles: ["dotenv-flow/config"],
   },
 });
