+++ packages/commons-test/package.json
@@ -22,9 +22,11 @@
     "@pagopa/eslint-config": "3.0.0",
     "@protobuf-ts/runtime": "2.9.4",
     "@testcontainers/postgresql": "10.2.2",
+    "@types/dotenv-flow": "3.2.0",
     "@types/jsonwebtoken": "9.0.6",
     "@types/uuid": "9.0.8",
     "@zodios/core": "10.9.6",
+    "dotenv-flow": "3.2.0",
     "axios": "1.6.8",
     "jsonwebtoken": "9.0.2",
     "pagopa-interop-commons": "workspace:*",
@@ -34,7 +36,7 @@
     "ts-pattern": "5.0.6",
     "typescript": "5.1.3",
     "uuid": "9.0.1",
-    "vitest": "0.33.0",
+    "vitest": "1.5.2",
     "zod": "3.22.3"
   }
 }
