+++ packages/commons-test/src/containerTestUtils.ts
@@ -1,5 +1,11 @@
-import { EventStoreConfig, ReadModelDbConfig } from "pagopa-interop-commons";
+import {
+  EventStoreConfig,
+  FileManagerConfig,
+  LoggerConfig,
+  ReadModelDbConfig,
+} from "pagopa-interop-commons";
 import { GenericContainer } from "testcontainers";
+import { z } from "zod";
 
 export const TEST_MONGO_DB_PORT = 27017;
 export const TEST_MONGO_DB_IMAGE = "mongo:4";
@@ -70,3 +76,14 @@ export const minioContainer = (config: {
       `mkdir -p /data/${config.s3Bucket} && /usr/bin/minio server /data`,
     ])
     .withExposedPorts(TEST_MINIO_PORT);
+
+export const TestContainersConfig = LoggerConfig.and(ReadModelDbConfig)
+  .and(EventStoreConfig)
+  .and(FileManagerConfig)
+  .and(
+    z
+      .object({ S3_BUCKET: z.string().optional() })
+      .transform((c) => ({ s3Bucket: c.S3_BUCKET }))
+  );
+
+export type TestContainersConfig = z.infer<typeof TestContainersConfig>;
