+++ packages/commons-test/src/index.ts
@@ -3,3 +3,5 @@ export * from "./eventStoreTestUtils.js";
 export * from "./testUtils.js";
 export * from "./containerTestUtils.js";
 export * from "./riskAnalysisTestUtils.js";
+export * from "./setupTestContainersVitest.js";
+export * from "./setupTestContainersVitestGlobal.js";
