+++ packages/event-migration/package.json
@@ -38,6 +38,6 @@
     "prettier": "2.8.8",
     "ts-node": "10.9.2",
     "typescript": "5.1.3",
-    "vitest": "0.33.0"
+    "vitest": "1.5.2"
   }
 }
