+++ packages/notifier-seeder/package.json
@@ -29,12 +29,11 @@
     "prettier": "2.8.8",
     "ts-node": "10.9.2",
     "typescript": "5.1.3",
-    "dotenv-flow": "3.2.0",
     "kafkajs": "2.2.4",
     "@types/uuid": "9.0.8",
     "mkdirp": "3.0.1",
     "testcontainers": "10.2.2",
-    "vitest": "0.33.0"
+    "vitest": "1.5.2"
   },
   "dependencies": {
     "@protobuf-ts/plugin": "2.9.4",
