+++ packages/purpose-process/package.json
@@ -29,7 +29,7 @@
     "testcontainers": "10.2.2",
     "ts-node": "10.9.2",
     "typescript": "5.1.3",
-    "vitest": "0.33.0"
+    "vitest": "1.5.2"
   },
   "dependencies": {
     "@zodios/core": "10.9.6",
