+++ packages/purpose-readmodel-writer/package.json
@@ -29,7 +29,7 @@
     "testcontainers": "10.8.1",
     "ts-node": "10.9.2",
     "typescript": "5.1.3",
-    "vitest": "0.33.0"
+    "vitest": "1.5.2"
   },
   "dependencies": {
     "@protobuf-ts/runtime": "2.9.4",
